Specify mock contract of ERC4626 in comments
Hello, I was interested in helping with this issue as a fairly new open-source contributor. I added a comment on line 2 that explicitly states this a mock contract implementation of ERC4626. I also edited the comment on line 21 to remind that this is a mock contract. Please let me know if this satisfies what you were looking for, or if you had something else in mind.